Rename the LTFS FreeBSD backend to conform to the LTFS backend naming conventions.
Old:  libtape-camtape-freebsd.so and tape_freebsd_camtape
New:  libtape-cam.so and tape_freebsd_cam

As Abe-san explained:

"We introduced weak convention against tape backend name (Sorry it
is undocumented yet...).

Naming of the backend is tape_[platform]_[accessdriver]_[supportdrive]
Naming of the library file is libtape-[accessdriver]-[supportdrive]

For example, ibmtape support on sg (on linux) is

Platform: linux
Access driver: sg
Support drives: ibmtape"

He also explained that the drive support portion is implicit and
can be omitted.

So, for now, we'll call it the CAM backend.  It uses both the sa(4)
and pass(4) drivers.  It is possible that at some point, we'll do
a passthrough-only FreeBSD backend, like the Linux sg backend.  If
we do that, it will probably be best to do that by modifying the sg
backend to talk to a generic SCSI passthrough layer that works with
Linux, FreeBSD and other OSes.  If we do that, we'll probably need
to consider renaming the sg backend to something else at that
point.

I left the function names and other variables in the backend as
camtape_*.  Since there are lots of functions in libcam on FreeBSD
that are prefixed with cam_, renaming the functions inside the
backend to cam_* may produce conflicts.  For instance, the LTFS
backend has a camtape_send_ccb(), and libcam has a cam_send_ccb().
Renaming the backend function to cam_send_ccb() would produce an
obvious conflict.

conf/Makefile.am:
	libtape-camtape-freebsd.so -> libtape-cam.so

configure.ac:
	DEFAULT_TAPE=camtape -> DEFAULT_TAPE=cam

messages/Makefile.am:
	tape_freebsd_camtape_dat.o -> tape_freebsd_cam_dat.o

messages/tape_freebsd_camtape -> messages/tape_freebsd_cam:

src/tape_drivers/freebsd/cam/Makefile.am:
	libtape_camtape_freebsd -> libtape_cam
	camtape_cmn.c -> cam_cmn.c
	camtape_tc.c -> cam_tc.c

src/tape_drivers/freebsd/cam/camtape_cmn.c:
	Renamed to cam_cmn.c.

src/tape_drivers/freebsd/cam/camtape_cmn.h:
	Renamed to cam_cmn.h.

src/tape_drivers/freebsd/cam/camtape_tc.c:
	Renamed to cam_tc.c.

	Change the include for camtape_cmn.h to cam_cmn.h.

	The message data bundle is now called tape_freebsd_cam_dat.
